requirements.txt
fastapi
uvicorn
PyPDF2
python-docx
python-dotenv
openai
python-dotenv